+++ packages/form/src/components/Dependency/index.tsx
@@ -38,7 +38,7 @@ const ProFormDependency: React.FC<ProFormDependencyProps> = ({
       if (formListField.name === undefined || ignoreFormListField) {
         return [itemName].flat(1) as string[];
       }
-      return [formListField.listName, itemName].flat(1) as string[];
+      return [formListField.listName, itemName].flat(2) as string[];
     },
     [formListField.listName, formListField.name, ignoreFormListField],
   );
@@ -64,7 +64,13 @@ const ProFormDependency: React.FC<ProFormDependencyProps> = ({
       }}
     >
       {(form) => {
-        const values = form.getFieldsValue(names);
+        const values = names.reduce((pre, next) => {
+          const value = form.getFieldsValue([next].flat(1));
+          return {
+            ...pre,
+            ...value,
+          };
+        }, {});
         const nameValues = name
           .map((itemName) => {
             const namePath = getNamePath(itemName);