+++ packages/form/src/layouts/ProForm/index.tsx
@@ -4,6 +4,7 @@ import { Form } from 'antd';
 import Group from '../../components/Group';
 import type { CommonFormProps } from '../../BaseForm';
 import BaseForm from '../../BaseForm';
+import ProFormItem from '../../components/FormItem';
 
 export type ProFormProps<T = Record<string, any>> = Omit<FormProps<T>, 'onFinish'> &
   CommonFormProps<T>;
@@ -31,6 +32,6 @@ function ProForm<T = Record<string, any>>(props: ProFormProps<T>) {
 
 ProForm.Group = Group;
 ProForm.useForm = Form.useForm;
-ProForm.Item = Form.Item;
+ProForm.Item = ProFormItem;
 
 export default ProForm;